Changed some class names, refactored packages
In order to go MVVM I started separating the three elements in separate
packages even though I know it's not a good architecture idea but this
is a study project, so I think it's good to make MVVM explicit.

First step is to move model classes to the model package. I also renamed
the files and classes to make them easier to understand and created an
interface to be used as a layer between the ViewModel and the Model, so
in the future any model implementing the very simple interface will be
a good source of data